Build
- Eliminated build warnings.

+++ src/TLSClient.cpp
@@ -193,8 +193,8 @@ void TLSClient::send (const std::string& data)
   packet[2] = l >>8;
   packet[3] = l;
 
-  int total = 0;
-  int remaining = packet.length ();
+  unsigned int total = 0;
+  unsigned int remaining = packet.length ();
 
   while (total < packet.length ())
   {
@@ -209,8 +209,8 @@ void TLSClient::send (const std::string& data)
     if (status == -1)
       break;
 
-    total     += status;
-    remaining -= status;
+    total     += (unsigned int) status;
+    remaining -= (unsigned int) status;
   }
 
   if (_debug)
